+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0:
+
+- `useMemoized` callback doesn't take the previous value anymore (to match React API)
+Use `useValueChanged` instead.
+- Introduced `useEffect` and `useStreamController`
+- fixed a bug where hot-reload while reordering/adding hooks did not work properly
+- improved readme
+
 ## 0.0.1:
 
 Added a few common hooks: